backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-docx
httpx
